+++ de.metas.adempiere.libero.libero/src/main/java/org/eevolution/event/MaterialDocumentListener.java
@@ -8,7 +8,9 @@ import java.util.Date;
 import org.adempiere.ad.trx.api.ITrxManager;
 import org.adempiere.util.Services;
 import org.adempiere.util.lang.Mutable;
+import org.eevolution.model.I_DD_Order;
 import org.eevolution.model.I_PP_Order;
+import org.eevolution.mrp.spi.impl.ddorder.DDOrderProducer;
 import org.eevolution.mrp.spi.impl.pporder.PPOrderProducer;
 import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
@@ -17,10 +19,12 @@ import com.google.common.annotations.VisibleForTesting;
 
 import de.metas.document.engine.IDocActionBL;
 import de.metas.logging.LogManager;
+import de.metas.material.event.DDOrderRequestedEvent;
 import de.metas.material.event.MaterialDemandEvent;
 import de.metas.material.event.MaterialEvent;
 import de.metas.material.event.MaterialEventListener;
 import de.metas.material.event.PPOrderRequestedEvent;
+import de.metas.material.event.ddorder.DDOrder;
 import de.metas.material.event.pporder.PPOrder;
 import lombok.NonNull;
 
@@ -58,24 +62,38 @@ public class MaterialDocumentListener implements MaterialEventListener
 
 	private final PPOrderProducer ppOrderProducer;
 
-	public MaterialDocumentListener(@NonNull final PPOrderProducer ppOrderProducer)
+	private final DDOrderProducer ddOrderProducer;
+
+	public MaterialDocumentListener(
+			@NonNull final PPOrderProducer ppOrderProducer,
+			@NonNull final DDOrderProducer ddOrderProducer)
 	{
 		this.ppOrderProducer = ppOrderProducer;
+		this.ddOrderProducer = ddOrderProducer;
 	}
 
 	@Override
 	public void onEvent(final MaterialEvent event)
 	{
-		if (!(event instanceof PPOrderRequestedEvent))
+		if (event instanceof PPOrderRequestedEvent)
 		{
-			return;
-		}
-		logger.info("Received event {}", event);
+			logger.info("Received event {}", event);
 
-		final PPOrderRequestedEvent productionOrderEvent = (PPOrderRequestedEvent)event;
-		createProductionOrderInTrx(
-				productionOrderEvent.getPpOrder(),
-				Date.from(productionOrderEvent.getEventDescr().getWhen()));
+			final PPOrderRequestedEvent productionOrderEvent = (PPOrderRequestedEvent)event;
+			createProductionOrderInTrx(
+					productionOrderEvent.getPpOrder(),
+					Date.from(productionOrderEvent.getEventDescr().getWhen()));
+		}
+		else if (event instanceof DDOrderRequestedEvent)
+		{
+			logger.info("Received event {}", event);
+
+			final DDOrderRequestedEvent distributionOrderEvent = (DDOrderRequestedEvent)event;
+			createDistributionOrderInTrx(
+					distributionOrderEvent.getDdOrder(),
+					Date.from(distributionOrderEvent.getEventDescr().getWhen()));
+		}
+
 	}
 
 	private I_PP_Order createProductionOrderInTrx(
@@ -110,5 +128,32 @@ public class MaterialDocumentListener implements MaterialEventListener
 			Services.get(IDocActionBL.class).processEx(ppOrderRecord, ACTION_Complete, STATUS_Completed);
 		}
 		return ppOrderRecord;
+	}
+
+	private I_DD_Order createDistributionOrderInTrx(
+			@NonNull final DDOrder ddOrder,
+			@NonNull final Date dateOrdered)
+	{
+		final Mutable<I_DD_Order> result = new Mutable<>();
+
+		final ITrxManager trxManager = Services.get(ITrxManager.class);
+		trxManager
+				.run(trxName -> {
+					result.setValue(createDistributionOrder(ddOrder, dateOrdered));
+				});
+		return result.getValue();
+	}
+
+	@VisibleForTesting
+	I_DD_Order createDistributionOrder(
+			@NonNull final DDOrder ddOrder,
+			@NonNull final Date dateOrdered)
+	{
+		final I_DD_Order ddOrderRecord = ddOrderProducer.createDDOrder(ddOrder, dateOrdered);
+		if (ddOrderRecord.getPP_Product_Planning().isDocComplete())
+		{
+			Services.get(IDocActionBL.class).processEx(ddOrderRecord, ACTION_Complete, STATUS_Completed);
+		}
+		return ddOrderRecord;
 	}
 }