+++ de.metas.material/dispo-service/src/main/sql/postgresql/system/5461100_sys_gh523-metasfresh_augment_table_MD_Candidate_Prod_Detail.sql
@@ -1,6 +1,6 @@
 -- 2017-04-30T09:32:02.755
 -- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
-INSERT INTO AD_Column (AD_Client_ID,AD_Column_ID,AD_Element_ID,AD_Org_ID,AD_Reference_ID,AD_Table_ID,AllowZoomTo,ColumnName,Created,CreatedBy,DDL_NoForeignKey,EntityType,FieldLength,IsActive,IsAdvancedText,IsAllowLogging,IsAlwaysUpdateable,IsAutocomplete,IsCalculated,IsDimension,IsDLMPartitionBoundary,IsEncrypted,IsGenericZoomKeyColumn,IsGenericZoomOrigin,IsIdentifier,IsKey,IsLazyLoading,IsMandatory,IsParent,IsSelectionColumn,IsStaleable,IsSyncDatabase,IsTranslated,IsUpdateable,IsUseDocSequence,Name,SeqNo,Updated,UpdatedBy,Version) VALUES (0,556534,53276,0,30,540810,'N','PP_Order_ID',TO_TIMESTAMP('2017-04-30 09:32:02','YYYY-MM-DD HH24:MI:SS'),100,'N','de.metas.material.dispo.service',10,'Y','N','Y','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','Y','N','Produktionsauftrag',0,TO_TIMESTAMP('2017-04-30 09:32:02','YYYY-MM-DD HH24:MI:SS'),100,0)
+INSERT INTO AD_Column (AD_Client_ID,AD_Column_ID,AD_Element_ID,AD_Org_ID,AD_Reference_ID,AD_Table_ID,AllowZoomTo,ColumnName,Created,CreatedBy,DDL_NoForeignKey,EntityType,FieldLength,IsActive,IsAdvancedText,IsAllowLogging,IsAlwaysUpdateable,IsAutocomplete,IsCalculated,IsDimension,IsDLMPartitionBoundary,IsEncrypted,IsGenericZoomKeyColumn,IsGenericZoomOrigin,IsIdentifier,IsKey,IsLazyLoading,IsMandatory,IsParent,IsSelectionColumn,IsStaleable,IsSyncDatabase,IsTranslated,IsUpdateable,IsUseDocSequence,Name,SeqNo,Updated,UpdatedBy,Version) VALUES (0,556534,53276,0,30,540810,'N','PP_Order_ID',TO_TIMESTAMP('2017-04-30 09:32:02','YYYY-MM-DD HH24:MI:SS'),100,'N','de.metas.material.dispo',10,'Y','N','Y','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','Y','N','Produktionsauftrag',0,TO_TIMESTAMP('2017-04-30 09:32:02','YYYY-MM-DD HH24:MI:SS'),100,0)
 ;
 
 -- 2017-04-30T09:32:02.765
@@ -20,7 +20,7 @@ ALTER TABLE MD_Candidate_Prod_Detail ADD CONSTRAINT PPOrder_MDCandidateProdDetai
 
 -- 2017-04-30T09:32:26.273
 -- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
-INSERT INTO AD_Column (AD_Client_ID,AD_Column_ID,AD_Element_ID,AD_Org_ID,AD_Reference_ID,AD_Table_ID,AllowZoomTo,ColumnName,Created,CreatedBy,DDL_NoForeignKey,EntityType,FieldLength,IsActive,IsAdvancedText,IsAllowLogging,IsAlwaysUpdateable,IsAutocomplete,IsCalculated,IsDimension,IsDLMPartitionBoundary,IsEncrypted,IsGenericZoomKeyColumn,IsGenericZoomOrigin,IsIdentifier,IsKey,IsLazyLoading,IsMandatory,IsParent,IsSelectionColumn,IsStaleable,IsSyncDatabase,IsTranslated,IsUpdateable,IsUseDocSequence,Name,SeqNo,Updated,UpdatedBy,Version) VALUES (0,556535,53275,0,30,540810,'N','PP_Order_BOMLine_ID',TO_TIMESTAMP('2017-04-30 09:32:26','YYYY-MM-DD HH24:MI:SS'),100,'N','de.metas.material.dispo.service',10,'Y','N','Y','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','Y','N','Manufacturing Order BOM Line',0,TO_TIMESTAMP('2017-04-30 09:32:26','YYYY-MM-DD HH24:MI:SS'),100,0)
+INSERT INTO AD_Column (AD_Client_ID,AD_Column_ID,AD_Element_ID,AD_Org_ID,AD_Reference_ID,AD_Table_ID,AllowZoomTo,ColumnName,Created,CreatedBy,DDL_NoForeignKey,EntityType,FieldLength,IsActive,IsAdvancedText,IsAllowLogging,IsAlwaysUpdateable,IsAutocomplete,IsCalculated,IsDimension,IsDLMPartitionBoundary,IsEncrypted,IsGenericZoomKeyColumn,IsGenericZoomOrigin,IsIdentifier,IsKey,IsLazyLoading,IsMandatory,IsParent,IsSelectionColumn,IsStaleable,IsSyncDatabase,IsTranslated,IsUpdateable,IsUseDocSequence,Name,SeqNo,Updated,UpdatedBy,Version) VALUES (0,556535,53275,0,30,540810,'N','PP_Order_BOMLine_ID',TO_TIMESTAMP('2017-04-30 09:32:26','YYYY-MM-DD HH24:MI:SS'),100,'N','de.metas.material.dispo',10,'Y','N','Y','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','Y','N','Manufacturing Order BOM Line',0,TO_TIMESTAMP('2017-04-30 09:32:26','YYYY-MM-DD HH24:MI:SS'),100,0)
 ;
 
 -- 2017-04-30T09:32:26.275
@@ -40,7 +40,7 @@ ALTER TABLE MD_Candidate_Prod_Detail ADD CONSTRAINT PPOrderBOMLine_MDCandidatePr
 
 -- 2017-04-30T09:34:24.964
 -- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
-INSERT INTO AD_Element (AD_Client_ID,AD_Element_ID,AD_Org_ID,ColumnName,Created,CreatedBy,EntityType,IsActive,Name,PrintName,Updated,UpdatedBy) VALUES (0,543321,0,'PP_Order_DocStatus',TO_TIMESTAMP('2017-04-30 09:34:24','YYYY-MM-DD HH24:MI:SS'),100,'de.metas.material.dispo.service','Y','Belegstatus','Belegstatus',TO_TIMESTAMP('2017-04-30 09:34:24','YYYY-MM-DD HH24:MI:SS'),100)
+INSERT INTO AD_Element (AD_Client_ID,AD_Element_ID,AD_Org_ID,ColumnName,Created,CreatedBy,EntityType,IsActive,Name,PrintName,Updated,UpdatedBy) VALUES (0,543321,0,'PP_Order_DocStatus',TO_TIMESTAMP('2017-04-30 09:34:24','YYYY-MM-DD HH24:MI:SS'),100,'de.metas.material.dispo','Y','Belegstatus','Belegstatus',TO_TIMESTAMP('2017-04-30 09:34:24','YYYY-MM-DD HH24:MI:SS'),100)
 ;
 
 -- 2017-04-30T09:34:24.971
@@ -50,7 +50,7 @@ INSERT INTO AD_Element_Trl (AD_Language,AD_Element_ID, Description,Help,Name,PO_
 
 -- 2017-04-30T09:35:32.155
 -- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
-INSERT INTO AD_Column (AD_Client_ID,AD_Column_ID,AD_Element_ID,AD_Org_ID,AD_Reference_ID,AD_Reference_Value_ID,AD_Table_ID,AllowZoomTo,ColumnName,Created,CreatedBy,DDL_NoForeignKey,EntityType,FieldLength,IsActive,IsAdvancedText,IsAllowLogging,IsAlwaysUpdateable,IsAutocomplete,IsCalculated,IsDimension,IsDLMPartitionBoundary,IsEncrypted,IsGenericZoomKeyColumn,IsGenericZoomOrigin,IsIdentifier,IsKey,IsLazyLoading,IsMandatory,IsParent,IsSelectionColumn,IsStaleable,IsSyncDatabase,IsTranslated,IsUpdateable,IsUseDocSequence,Name,SeqNo,Updated,UpdatedBy,Version) VALUES (0,556536,543321,0,17,131,540810,'N','PP_Order_DocStatus',TO_TIMESTAMP('2017-04-30 09:35:32','YYYY-MM-DD HH24:MI:SS'),100,'N','de.metas.material.dispo.service',2,'Y','N','Y','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','Y','N','Belegstatus',0,TO_TIMESTAMP('2017-04-30 09:35:32','YYYY-MM-DD HH24:MI:SS'),100,0)
+INSERT INTO AD_Column (AD_Client_ID,AD_Column_ID,AD_Element_ID,AD_Org_ID,AD_Reference_ID,AD_Reference_Value_ID,AD_Table_ID,AllowZoomTo,ColumnName,Created,CreatedBy,DDL_NoForeignKey,EntityType,FieldLength,IsActive,IsAdvancedText,IsAllowLogging,IsAlwaysUpdateable,IsAutocomplete,IsCalculated,IsDimension,IsDLMPartitionBoundary,IsEncrypted,IsGenericZoomKeyColumn,IsGenericZoomOrigin,IsIdentifier,IsKey,IsLazyLoading,IsMandatory,IsParent,IsSelectionColumn,IsStaleable,IsSyncDatabase,IsTranslated,IsUpdateable,IsUseDocSequence,Name,SeqNo,Updated,UpdatedBy,Version) VALUES (0,556536,543321,0,17,131,540810,'N','PP_Order_DocStatus',TO_TIMESTAMP('2017-04-30 09:35:32','YYYY-MM-DD HH24:MI:SS'),100,'N','de.metas.material.dispo',2,'Y','N','Y','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','N','Y','N','Belegstatus',0,TO_TIMESTAMP('2017-04-30 09:35:32','YYYY-MM-DD HH24:MI:SS'),100,0)
 ;
 
 -- 2017-04-30T09:35:32.157