+++ de.metas.scripts.rollout/src/main/bash/tools.sh
@@ -129,6 +129,8 @@ check_java_version()
             if [[ ! $JAVA_VERSION == "8" ]]; then
                 trace check_java_version "[ERROR] JAVA-Version does not match supported version (found Java 1.${JAVA_VERSION})! Only Java 1.8 JDK supported! Check ' http://docs.metasfresh.org/howto_collection/Wie_aktualisiere_ich_die_Java_Version_auf_meinem_server.html ' for more infos and make sure 'java -version' is 1.8.x ! NOTE: you may also need to update to Java-8 on your Clients: ' http://docs.metasfresh.org/howto_collection/Wie_aktualisiere_ich_die_Java_Version_auf_meinem_client.html '"
                 exit 1
+			else
+				trace check_java_version "OK: JAVA_VERSION=${JAVA_VERSION}"
             fi
         else
             trace check_java_version "could not find valid /usr/bin/java. assuming Java 1.8 JDK is installed."